Rename NetWeightsPanel to WeightsPanel and remove edit functionality.
